properly resetting encoder values
git-svn-id: svn://localhost:5050/magic2010/trunk@47 61389559-6c2a-4883-acdc-2d27d9f19ada

+++ visPlugins/VehicleDynamics2D.hh
@@ -42,7 +42,7 @@ namespace Magic
     
     private: double vDes,wDes;
     
-    private: int16_t fr, fl, rr, rl;
+    private: double fr, fl, rr, rl;
     private: uint16_t cntr;
     private: double vmin,vmax,wmin,wmax;
   };
